% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mrds_GOF.R
\name{get_detection}
\alias{get_detection}
\title{Utility function to get a detection history given a multinomial cell}
\usage{
get_detection(cur.bin, n.obs.bins)
}
\arguments{
\item{cur.bin}{Multinomial cell bin number (1:(n.obs.bins^2+2*n.obs.bins))}

\item{n.obs.bins}{Number of observable distance bins (n_S in paper)}
}
\value{
A row and column entry
}
\description{
Utility function to get a detection history given a multinomial cell
}
\author{
Paul B. Conn
}
