% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mrds_GOF.R
\name{mat_sample}
\alias{mat_sample}
\title{Utility function to randomly sample a single cell of a matrix when entries of the matrix are probability of selection}
\usage{
mat_sample(pmatrix)
}
\arguments{
\item{pmatrix}{Matrix of probabilities}
}
\value{
A row and column entry
}
\description{
Utility function to randomly sample a single cell of a matrix when entries of the matrix are probability of selection
}
\author{
Paul B. Conn
}
