% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRDSmove_IntLik_ObsDep.R
\name{MRDSmove_IntLik_ObsDep}
\alias{MRDSmove_IntLik_ObsDep}
\title{Negative log likelihood function for MRDS double observer data subject to movement and measurement error for models estimating observer dependence}
\usage{
MRDSmove_IntLik_ObsDep(Par, Data, p.formula, dep.formula, Bin.widths, Obs.bins,
  Move.fix = NULL, gaussian.move = FALSE, gaussian.meas = FALSE)
}
\arguments{
\item{Par}{Parameter vector, including detection parameters, movement error SDs (left and right tail), measurement error SD}

\item{Data}{A design.matrix with the following column names: "match" indicates which records match with which (there should be two records
 for each detection, one for each observer), "observer","species" (provides species or other grouping variable: abundance estimates will be provided separately for each), 
"obs.dist" (observed distance; NA if missing), "g_size" (group/cluster size), "moving" (binary indicator for moving/not moving), 
"detected" (binary detection/nondetection). Finally, the optional column "count" gives the count of histories observed a particular type (to make likelihood calculations faster)
 Additional covariates may also be provided and used in formula (e.g. "other"; see below)}

\item{p.formula}{Formula object giving formula for detection probability for each observer.  Variables can be linked to column names in Data, 
but need to use "distance" to represent distance and "distance2" to represent distance^2.}

\item{dep.formula}{Formula object for implementing observer dependence as described in discussion of MacKenzie and Clement 2016.  Note that the
formula must use "det_other" in some fashion; for instance, an interaction between "det_other" and "distance" (i.e. ~0+det_other:distance) can be used to implement point indendence.}

\item{Bin.widths}{Vector of distance bin widths}

\item{Obs.bins}{A vector giving which bins are observed (e.g. 1:3 if bins 1-3 are observed)}

\item{Move.fix}{A indicator vector giving which movement/measurement error parameters to fix to 0 (omit if all are estimated)}

\item{gaussian.move}{If TRUE, uses two half-normals for movement; if FALSE (default), uses exponential / half exponential (Laplace dist)}

\item{gaussian.meas}{If TRUE, uses normal distribution for measurement error; if FALSE (default), uses double exponential (Laplace dist)}
}
\value{
Negative log-likelihood function
}
\description{
Negative log likelihood function for MRDS double observer data subject to movement and measurement error for models estimating observer dependence
}
\author{
Paul B. Conn
}
\keyword{mrds}
\keyword{simulation,}
