% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mrds_GOF_ObsDep.R
\name{sim_mrds_GOF_ObsDep}
\alias{sim_mrds_GOF_ObsDep}
\title{function to simulate MRDS data w/ movement and measurement error from model estimates.  Note: not set up to deal with counts>1}
\usage{
sim_mrds_GOF_ObsDep(Par, Data, p.formula, dep.formula, Bin.widths, Obs.bins,
  Move.fix = Move.fix, gaussian.move = FALSE, gaussian.meas = FALSE)
}
\arguments{
\item{Par}{Parameter vector, including detection parameters, movement error SDs (left and right tail), measurement error SD}

\item{Data}{A design.matrix with the following column names: "match" indicates which records match with which (there should be two records
 for each detection, one for each observer), "observer","species" (provides species or other grouping variable: abundance estimates will be provided separately for each), 
"obs.dist" (observed distance; NA if missing), "g_size" (group/cluster size), "moving" (binary indicator for moving/not moving), 
"detected" (binary detection/nondetection). Finally, the optional column "count" gives the count of histories observed a particular type (to make likelihood calculations faster).
 Additional covariates may also be provided and used in formula (e.g. "other"; see below)}

\item{p.formula}{Formula object giving formula for detection probability for each observer.  Variables can be linked to column names in Data, 
but need to use "distance" to represent distance and "distance2" to represent distance^2.}

\item{dep.formula}{Another formula object, this time giving the structure of observer dependence.  For example, ~0+distance:det_other can provide
a model for point independence sensu MacKenzie & Clement (2016)}

\item{Bin.widths}{Vector of distance bin widths}

\item{Obs.bins}{A vector giving which bins are observed (e.g. 1:3 if bins 1-3 are observed)}

\item{Move.fix}{A indicator vector giving which movement/measurement error parameters to fix to 0 (omit if all are estimated)}

\item{gaussian.move}{If TRUE, uses two half-normals for movement; if FALSE (default), uses exponential / half exponential (Laplace dist)}

\item{gaussian.meas}{If TRUE, uses normal distribution for measurement error; if FALSE (default), uses double exponential (Laplace dist)}
}
\value{
a community mrds dataset
}
\description{
function to simulate MRDS data w/ movement and measurement error from model estimates.  Note: not set up to deal with counts>1
}
\author{
Paul B. Conn
}
\keyword{mrds}
\keyword{simulation,}
