% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mrds.R
\name{simulate_mrds}
\alias{simulate_mrds}
\title{simulate waterfowl community MRDS double observer data subject to movement and measurement error}
\usage{
simulate_mrds(n_species, n_bins, n_obs_bins, measure_par = 0.6,
  move_par = c(0.1, 1.2), seed = NULL, gaussian = FALSE,
  point_indep = FALSE)
}
\arguments{
\item{n_species}{Number of species to simulate data for}

\item{n_bins}{Number of bins for complete data (note that measurement error can result in some of these being detected)}

\item{n_obs_bins}{Number of bins observed}

\item{measure_par}{Gives measurement error sd (if gaussian = TRUE) or measurement error rate (if gaussian=FALSE)}

\item{move_par}{2 parameter vector movement sd (if gaussian = TRUE) or rate (if gaussian=FALSE) for responseive movement}

\item{seed}{Optional; set the random # seed}

\item{gaussian}{If TRUE, uses a Gaussian distribution for measurement error and a half-normal for movement; if FALSE (default), uses exponential / half exponential (Laplace dist)}

\item{point_indep}{if TRUE (default is FALSE), implements point independence via a random effect}
}
\value{
a community mrds dataset
}
\description{
simulate waterfowl community MRDS double observer data subject to movement and measurement error
}
\author{
Paul B. Conn
}
\keyword{mrds}
\keyword{simulation,}
