\docType{data}
\name{waterfowl}
\alias{waterfowl}
\title{MRDS data for waterfowl surveyed in northern Canada}
\description{
  A data.frame consisting of the following columns
  1) Species
  2) CH - capture history for each observation (detection for each observer; note a '1' represents the '01' distory
  3) count - a field of all ones
  4) groupFLY - an indicator for whether the bird group was flying or not
  5) GROUPSZ - the number of birds detected (e.g. flock size) 
  6) DISTFRONT - the distance bin recorded by the front observer
  7) DISTBACK - distance bin recorded by the back observer
}
\author{
  Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{waterfowl}

